[perf] powerpc: Handle events that raise an exception without overflowing (Steve Best) [767917 755737] {CVE-2011-4611}

+++ arch/powerpc/kernel/perf_event.c
@@ -1270,7 +1270,7 @@ unsigned long perf_instruction_pointer(struct pt_regs *regs)
 
 static bool pmc_overflow(unsigned long val)
 {
-	if ((unsigned long)val < 0)
+	if ((int)val < 0)
 		return true;
 
 	/*
